Backend/requirements.txt
Flask==3.0.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Werkzeug==3.0.3
scikit-learn==1.5.2
joblib==1.4.2
numpy>=1.26.0
paho-mqtt==1.6.1
pandas>=2.0.0
openpyxl>=3.1.0